TEST-355 MicroprofileSecureEndpointTest Failures on Payara Micro
