+++ samples/microprofile-endpoints/insecure/pom.xml
@@ -31,18 +31,13 @@
 
     <parent>
         <groupId>fish.payara.samples</groupId>
-        <artifactId>samples-parent</artifactId>
+        <artifactId>microprofile-endpoints</artifactId>
         <version>1.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>microprofile-endpoints</artifactId>
-    <name>Payara Samples - Payara - Microprofile endpoints</name>
-    <packaging>pom</packaging>
-
-    <modules>
-        <module>secure</module>
-        <module>insecure</module>
-    </modules>
+    <artifactId>microprofile-insecure-endpoints</artifactId>
+    <name>Payara Samples - Payara - Microprofile insecure endpoints</name>
+    <packaging>war</packaging>
 
     <dependencies>
         <dependency>
@@ -52,4 +47,21 @@
         </dependency>
     </dependencies>
 
+    <profiles>
+        <profile>
+            <id>payara-micro-managed</id>
+            <build>
+                <plugins>
+                    <plugin>
+                        <artifactId>maven-surefire-plugin</artifactId>
+                        <configuration>
+                            <environmentVariables>
+                                <EXTRA_MICRO_OPTIONS>--postbootcommandfile ${basedir}\src\test\resources\post-boot-commands.txt --sslPort 8181 --autoBindSsl</EXTRA_MICRO_OPTIONS>
+                            </environmentVariables>
+                        </configuration>
+                    </plugin>
+                </plugins>
+            </build>
+        </profile>
+    </profiles>
 </project>
