+++ samples/microprofile-endpoints/insecure/src/test/resources/post-boot-commands.txt
@@ -1,0 +1,2 @@
+set-metrics-configuration --endpoint=mpmetrics
+set-microprofile-healthcheck-configuration --endpoint=mphealth
