+++ samples/microprofile-endpoints/secure/pom.xml
@@ -31,25 +31,30 @@
 
     <parent>
         <groupId>fish.payara.samples</groupId>
-        <artifactId>samples-parent</artifactId>
+        <artifactId>microprofile-endpoints</artifactId>
         <version>1.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>microprofile-endpoints</artifactId>
-    <name>Payara Samples - Payara - Microprofile endpoints</name>
-    <packaging>pom</packaging>
+    <artifactId>microprofile-secure-endpoints</artifactId>
+    <name>Payara Samples - Payara - Microprofile Secure endpoints</name>
+    <packaging>war</packaging>
 
-    <modules>
-        <module>secure</module>
-        <module>insecure</module>
-    </modules>
-
-    <dependencies>
-        <dependency>
-            <groupId>${project.groupId}</groupId>
-            <artifactId>test-utils</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-    </dependencies>
-
+    <profiles>
+        <profile>
+            <id>payara-micro-managed</id>
+            <build>
+                <plugins>
+                    <plugin>
+                        <artifactId>maven-surefire-plugin</artifactId>
+                        <configuration>
+                            <skipTests>${skipClientCertificate}</skipTests>
+                            <environmentVariables>
+                                <EXTRA_MICRO_OPTIONS>--postbootcommandfile ${basedir}\src\test\resources\post-boot-commands.txt --sslPort 8181 --autoBindSsl</EXTRA_MICRO_OPTIONS>
+                            </environmentVariables>
+                        </configuration>
+                    </plugin>
+                </plugins>
+            </build>
+        </profile>
+    </profiles>
 </project>
