+++ samples/openid/src/test/java/fish/payara/security/oidc/test/UseCookiesTest.java
@@ -41,14 +41,9 @@ package fish.payara.security.oidc.test;
 
 import com.gargoylesoftware.htmlunit.WebClient;
 import static fish.payara.security.annotations.OpenIdAuthenticationDefinition.OPENID_MP_USE_SESSION;
-import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.net.URL;
-import java.security.KeyPair;
-import java.security.Security;
-import java.security.cert.X509Certificate;
-import org.bouncycastle.jce.provider.BouncyCastleProvider;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.container.test.api.OperateOnDeployment;
 import org.jboss.arquillian.container.test.api.RunAsClient;
@@ -59,18 +54,7 @@ import org.jboss.shrinkwrap.api.spec.WebArchive;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import static org.omnifaces.utils.security.Certificates.createTempJKSKeyStore;
 import fish.payara.samples.ServerOperations;
-import static org.omnifaces.utils.Lang.isEmpty;
-import static org.omnifaces.utils.security.Certificates.createTempJKSKeyStore;
-import static org.omnifaces.utils.security.Certificates.createTempJKSTrustStore;
-import static org.omnifaces.utils.security.Certificates.generateRandomRSAKeys;
-import static org.omnifaces.utils.security.Certificates.getCertificateChainFromServer;
-import static fish.payara.samples.ServerOperations.addCertificateToContainerTrustStore;
-import static fish.payara.samples.ServerOperations.addContainerSystemProperty;
-import static fish.payara.samples.ServerOperations.createSelfSignedCertificate;
-import static fish.payara.samples.ServerOperations.enableSSLDebug;
-import static fish.payara.samples.ServerOperations.getHostFromCertificate;
 
 /**
  *
@@ -96,42 +80,7 @@ public class UseCookiesTest {
 
     @Deployment(name = "openid-client")
     public static WebArchive createClientDeployment() {
-        System.out.println("\n*********** DEPLOYMENT START ***************************");
-
-        Security.addProvider(new BouncyCastleProvider());
-
-        // Enable to get detailed logging about the SSL handshake on the client
-        // For an explanation of the TLS handshake see: https://tls.ulfheim.net
-        if (System.getProperty("ssl.debug") != null) {
-            enableSSLDebug();
-        }
-
-        System.out.println("################################################################");
-
-        // ### Generate keys for the client, create a certificate, and add those to a new local key store
-        // Generate a Private/Public key pair for the client
-        KeyPair clientKeyPair = generateRandomRSAKeys();
-
-        // Create a certificate containing the client public key and signed with the private key
-        X509Certificate clientCertificate = createSelfSignedCertificate(clientKeyPair);
-
-        // Create a new local key store containing the client private key and the certificate
-        clientKeyStorePath = createTempJKSKeyStore(clientKeyPair.getPrivate(), clientCertificate);
-
-        // Enable to get detailed logging about the SSL handshake on the server
-        if (System.getProperty("ssl.debug") != null) {
-            System.out.println("Setting server SSL debug on");
-            addContainerSystemProperty("javax.net.debug", "ssl:handshake");
-        }
-
-        // Only test TLS v1.2 for now
-        System.setProperty("jdk.tls.client.protocols", "TLSv1.2");
-
-        // Add the client certificate that we just generated to the trust store of the server.
-        // That way the server will trust our certificate.
-        // Set the actual domain used with -Dpayara_domain=[domain name]
-        addCertificateToContainerTrustStore(clientCertificate);
-
+        clientKeyStorePath = ServerOperations.createClientKeyStore();
         StringAsset mpConfig = new StringAsset(OPENID_MP_USE_SESSION + "=" + Boolean.FALSE.toString());
         return OpenIdTestUtil
                 .createClientDeployment()
@@ -141,46 +90,7 @@ public class UseCookiesTest {
     @Before
     public void setup() throws FileNotFoundException, IOException {
         webClient = new WebClient();
-        baseHttps = ServerOperations.toContainerHttps(base);
-        if (baseHttps == null) {
-            throw new IllegalStateException("No https URL could be created from " + base);
-        }
-
-        // ### Ask the server for its certificate and add that to a new local trust store
-        // Server -> client : the trust store certificates are used to validate the certificate sent
-        // by the server
-        X509Certificate[] serverCertificateChain = getCertificateChainFromServer(baseHttps.getHost(), baseHttps.getPort());
-
-        if (!isEmpty(serverCertificateChain)) {
-
-            System.out.println("Obtained certificate from server. Storing it in client trust store");
-
-            String trustStorePath = createTempJKSTrustStore(serverCertificateChain);
-
-            System.out.println("Reading trust store from: " + trustStorePath);
-
-            webClient.getOptions().setSSLTrustStore(new File(trustStorePath).toURI().toURL(), "changeit", "jks");
-
-            // If the use.cnHost property is we try to extract the host from the server
-            // certificate and use exactly that host for our requests.
-            // This is needed if a server is listening to multiple host names, for instance
-            // localhost and example.com. If the certificate is for example.com, we can't
-            // localhost for the request, as that will not be accepted.
-            if (System.getProperty("use.cnHost") != null) {
-                System.out.println("use.cnHost set. Trying to grab CN from certificate and use as host for requests.");
-                baseHttps = getHostFromCertificate(serverCertificateChain, baseHttps);
-            }
-        } else {
-            System.out.println("Could not obtain certificates from server. Continuing without custom truststore");
-        }
-
-        System.out.println("Using client key store from: " + clientKeyStorePath);
-
-        // Client -> Server : the key store's private keys and certificates are used to sign
-        // and sent a reply to the server
-        webClient.getOptions().setSSLClientCertificate(new File(clientKeyStorePath).toURI().toURL(), "changeit", "jks");
-
-        System.out.println("*********** SETUP DONE ***************************\n");
+        baseHttps = ServerOperations.createClientTrustStore(webClient, base, clientKeyStorePath);
     }
 
     @Test
